feat: Add better user flow to remove reliance on serverProps
