+++ pages/home.tsx
@@ -3,7 +3,6 @@ import React from "react";
 import Content from "../src/components/content/Content";
 import MdLoader from "../src/components/md/MdLoader";
 import Home from "../src/components/Home";
-import UserNew from "../src/components/user/UserNew";
 import useFuegoAuth from "../src/fuego/useFuegoAuth";
 
 /**
@@ -26,7 +25,8 @@ function HomePage(): JSX.Element {
     }
     return <MdLoader />;
   } else if (!userValidating && userValid && userNoPreference) {
-    return <UserNew user={userValid} />;
+    router.push("/new");
+    return <MdLoader />;
   } else if (userValidHasPreference) {
     return (
       <Content title="Home">
