+++ pages/index.tsx
@@ -1,38 +1,20 @@
-import "firebase/auth";
-import type { GetServerSideProps, InferGetServerSidePropsType } from "next";
 import { useRouter } from "next/router";
-import { destroyCookie, parseCookies } from "nookies";
+import { parseCookies } from "nookies";
 import React from "react";
 import MdLoader from "../src/components/md/MdLoader";
-import UserNew from "../src/components/user/UserNew";
 import Welcome from "../src/components/Welcome";
 import useFuegoUser from "../src/fuego/useFuegoUser";
 
-export const getServerSideProps: GetServerSideProps = async (context) => {
-  const cookies = parseCookies(context);
-  return {
-    props: {
-      fuegoPending:
-        typeof cookies.fuegoPending !== "undefined"
-          ? cookies.fuegoPending
-          : false,
-      fuegoNewUser: typeof cookies.fuegoNewUser !== "undefined" ? true : false,
-    },
-  };
-};
-
-function App({
-  fuegoPending,
-  fuegoNewUser,
-}: InferGetServerSidePropsType<typeof getServerSideProps>): JSX.Element {
+export default function MainPage(): JSX.Element {
   const router = useRouter();
   const { user, isValidating } = useFuegoUser();
+  const cookies = parseCookies();
 
-  if ((fuegoPending && user === null) || isValidating) {
+  if (user === null || isValidating) {
     return <MdLoader />;
-  } else if ((fuegoPending || fuegoNewUser) && user) {
-    destroyCookie(null, "fuegoNewUser");
-    return <UserNew user={user} />;
+  } else if ((cookies?.fuegoPending || cookies?.fuegoNewUser) && user) {
+    router.push("/new");
+    return <MdLoader />;
   } else if (user) {
     router.push("/home");
     return <MdLoader />;
@@ -40,5 +22,3 @@ function App({
     return <Welcome />;
   }
 }
-
-export default App;
