+++ pages/settings.tsx
@@ -3,7 +3,6 @@ import React from "react";
 import Content from "../src/components/content/Content";
 import MdLoader from "../src/components/md/MdLoader";
 import Preferences from "../src/components/Settings";
-import NewUser from "../src/components/user/UserNew";
 import useFuegoAuth from "../src/fuego/useFuegoAuth";
 
 /**
@@ -26,7 +25,8 @@ function SettingsPage(): JSX.Element {
     }
     return <MdLoader />;
   } else if (!userValidating && userValid && userNoPreference) {
-    return <NewUser user={userValid} />;
+    router.push("/new");
+    return <MdLoader />;
   } else if (userValidHasPreference) {
     return (
       <Content title="Settings">
