+++ src/components/layouts/LayoutGradient.tsx
@@ -1,26 +1,10 @@
-import {
-  Box,
-  Button,
-  Flex,
-  Square,
-  Text,
-  useColorModeValue,
-  VStack,
-} from "@chakra-ui/react";
-import { setCookie } from "nookies";
-
-import Link from "next/link";
+import { Square } from "@chakra-ui/react";
+import type { PropsWithChildren } from "react";
 import React from "react";
-import fuego from "../fuego/fuego";
-import GoogleIcon from "./icons/GoogleIcon";
-import LogoIcon from "./icons/LogoIcon";
 
-interface Props {
-  type: "login" | "signup";
-}
-export default function Register({ type }: Props): JSX.Element {
-  const mdPurple = useColorModeValue("purple.700", "purple.200");
-  const typeText = type === "login" ? "Login in" : "Sign Up";
+export default function LayoutGradient({
+  children,
+}: PropsWithChildren<unknown>): JSX.Element {
   return (
     <Square
       h="100vh"
@@ -30,58 +14,7 @@ export default function Register({ type }: Props): JSX.Element {
         backgroundImage: `url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='402' height='335' viewBox='0 0 1080 900'%3E%3Cg fill-opacity='0.03'%3E%3Cpolygon fill='%23444' points='90 150 0 300 180 300'/%3E%3Cpolygon points='90 150 180 0 0 0'/%3E%3Cpolygon fill='%23AAA' points='270 150 360 0 180 0'/%3E%3Cpolygon fill='%23DDD' points='450 150 360 300 540 300'/%3E%3Cpolygon fill='%23999' points='450 150 540 0 360 0'/%3E%3Cpolygon points='630 150 540 300 720 300'/%3E%3Cpolygon fill='%23DDD' points='630 150 720 0 540 0'/%3E%3Cpolygon fill='%23444' points='810 150 720 300 900 300'/%3E%3Cpolygon fill='%23FFF' points='810 150 900 0 720 0'/%3E%3Cpolygon fill='%23DDD' points='990 150 900 300 1080 300'/%3E%3Cpolygon fill='%23444' points='990 150 1080 0 900 0'/%3E%3Cpolygon fill='%23DDD' points='90 450 0 600 180 600'/%3E%3Cpolygon points='90 450 180 300 0 300'/%3E%3Cpolygon fill='%23666' points='270 450 180 600 360 600'/%3E%3Cpolygon fill='%23AAA' points='270 450 360 300 180 300'/%3E%3Cpolygon fill='%23DDD' points='450 450 360 600 540 600'/%3E%3Cpolygon fill='%23999' points='450 450 540 300 360 300'/%3E%3Cpolygon fill='%23999' points='630 450 540 600 720 600'/%3E%3Cpolygon fill='%23FFF' points='630 450 720 300 540 300'/%3E%3Cpolygon points='810 450 720 600 900 600'/%3E%3Cpolygon fill='%23DDD' points='810 450 900 300 720 300'/%3E%3Cpolygon fill='%23AAA' points='990 450 900 600 1080 600'/%3E%3Cpolygon fill='%23444' points='990 450 1080 300 900 300'/%3E%3Cpolygon fill='%23222' points='90 750 0 900 180 900'/%3E%3Cpolygon points='270 750 180 900 360 900'/%3E%3Cpolygon fill='%23DDD' points='270 750 360 600 180 600'/%3E%3Cpolygon points='450 750 540 600 360 600'/%3E%3Cpolygon points='630 750 540 900 720 900'/%3E%3Cpolygon fill='%23444' points='630 750 720 600 540 600'/%3E%3Cpolygon fill='%23AAA' points='810 750 720 900 900 900'/%3E%3Cpolygon fill='%23666' points='810 750 900 600 720 600'/%3E%3Cpolygon fill='%23999' points='990 750 900 900 1080 900'/%3E%3Cpolygon fill='%23999' points='180 0 90 150 270 150'/%3E%3Cpolygon fill='%23444' points='360 0 270 150 450 150'/%3E%3Cpolygon fill='%23FFF' points='540 0 450 150 630 150'/%3E%3Cpolygon points='900 0 810 150 990 150'/%3E%3Cpolygon fill='%23222' points='0 300 -90 450 90 450'/%3E%3Cpolygon fill='%23FFF' points='0 300 90 150 -90 150'/%3E%3Cpolygon fill='%23FFF' points='180 300 90 450 270 450'/%3E%3Cpolygon fill='%23666' points='180 300 270 150 90 150'/%3E%3Cpolygon fill='%23222' points='360 300 270 450 450 450'/%3E%3Cpolygon fill='%23FFF' points='360 300 450 150 270 150'/%3E%3Cpolygon fill='%23444' points='540 300 450 450 630 450'/%3E%3Cpolygon fill='%23222' points='540 300 630 150 450 150'/%3E%3Cpolygon fill='%23AAA' points='720 300 630 450 810 450'/%3E%3Cpolygon fill='%23666' points='720 300 810 150 630 150'/%3E%3Cpolygon fill='%23FFF' points='900 300 810 450 990 450'/%3E%3Cpolygon fill='%23999' points='900 300 990 150 810 150'/%3E%3Cpolygon points='0 600 -90 750 90 750'/%3E%3Cpolygon fill='%23666' points='0 600 90 450 -90 450'/%3E%3Cpolygon fill='%23AAA' points='180 600 90 750 270 750'/%3E%3Cpolygon fill='%23444' points='180 600 270 450 90 450'/%3E%3Cpolygon fill='%23444' points='360 600 270 750 450 750'/%3E%3Cpolygon fill='%23999' points='360 600 450 450 270 450'/%3E%3Cpolygon fill='%23666' points='540 600 630 450 450 450'/%3E%3Cpolygon fill='%23222' points='720 600 630 750 810 750'/%3E%3Cpolygon fill='%23FFF' points='900 600 810 750 990 750'/%3E%3Cpolygon fill='%23222' points='900 600 990 450 810 450'/%3E%3Cpolygon fill='%23DDD' points='0 900 90 750 -90 750'/%3E%3Cpolygon fill='%23444' points='180 900 270 750 90 750'/%3E%3Cpolygon fill='%23FFF' points='360 900 450 750 270 750'/%3E%3Cpolygon fill='%23AAA' points='540 900 630 750 450 750'/%3E%3Cpolygon fill='%23FFF' points='720 900 810 750 630 750'/%3E%3Cpolygon fill='%23222' points='900 900 990 750 810 750'/%3E%3Cpolygon fill='%23222' points='1080 300 990 450 1170 450'/%3E%3Cpolygon fill='%23FFF' points='1080 300 1170 150 990 150'/%3E%3Cpolygon points='1080 600 990 750 1170 750'/%3E%3Cpolygon fill='%23666' points='1080 600 1170 450 990 450'/%3E%3Cpolygon fill='%23DDD' points='1080 900 1170 750 990 750'/%3E%3C/g%3E%3C/svg%3E")`,
       }}
     >
-      <Box
-        p="10"
-        border="1px solid"
-        borderColor="gray.500"
-        borderRadius="md"
-        bg="white"
-      >
-        <Box mb={10}>
-          <Flex align="center" justifyContent="center">
-            <LogoIcon boxSize={6} mr={1} color={mdPurple} />
-            <Link href="/">
-              <Text
-                fontSize={{ base: "md", md: "2xl" }}
-                color={mdPurple}
-                fontWeight="medium"
-                cursor="pointer"
-              >
-                mediaDiary
-              </Text>
-            </Link>
-          </Flex>
-        </Box>
-        <VStack align="stretch" spacing={4}>
-          <Button
-            onClick={handleLogin}
-            variant="outline"
-            leftIcon={<GoogleIcon />}
-          >
-            {typeText} with Google
-          </Button>
-        </VStack>
-        <Box textAlign="center" mt={10}>
-          <Text fontSize="xs" color="gray.400">
-            By continuing, you agree to mediaDiary
-          </Text>
-          <Text fontSize="xs" color="gray.400">
-            Terms of Service, Privacy Policy
-          </Text>
-        </Box>
-      </Box>
+      {children}
     </Square>
   );
-
-  function handleLogin() {
-    const provider = new fuego.auth.GoogleAuthProvider();
-    provider.setCustomParameters({ prompt: "select_account" });
-    fuego
-      .auth()
-      .setPersistence(fuego.auth.Auth.Persistence.LOCAL)
-      .then(() => {
-        setCookie(null, "fuegoPending", "true", {});
-        fuego.auth().signInWithRedirect(provider);
-      });
-  }
 }
